Server Auth Logging, Moved Server & Client Modules
